+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fix NPE in BankOrder generation on missing bank name.
 - Timesheets : use the timesheet user to filter the projects/tasks instead of the connected user.
 - Cannot create a new bank order from the menu entry "awaiting signature".
+- Birt Template : replace system char in fileName.
 
 ## [4.2.4] - 2018-07-12
 ### Improvements